+++ common/src/stack/command/stack/commands/add/pallet/imp_native_sles.py
@@ -5,8 +5,8 @@
 # @copyright@
 
 import os
-import sys
-import stack
+import shlex
+import subprocess
 import stack.commands
 from stack.exception import CommandError
 
@@ -33,18 +33,18 @@ class Implementation(stack.commands.Implementation):
 		# Clean out the existing roll directory if asked
 		#
 		if clean and os.path.exists(new_roll_dir):
-			print('Cleaning %s %s-%s' % (name, vers, release), end=' ')
-			print('for %s from the pallets directory' % arch)
+			self.owner.out.write('Cleaning %s %s-%s for %s from the pallets directory\n'
+				% (name, vers, release, arch))
 
-			if os.path.exists(new_roll_dir):
-				os.system('/bin/rm -rf %s' % new_roll_dir)
+			if not self.owner.dryrun:
+				if os.path.exists(new_roll_dir):
+					os.system('/bin/rm -rf %s' % new_roll_dir)
 
 		#
 		# copy the roll to the HD
 		#
-		sys.stdout.write('Copying %s %s-%s to pallets ...' %
+		self.owner.out.write('Copying %s %s-%s to pallets ...\n' %
 			(name, vers, release))
-		sys.stdout.flush()
 
 		os.chdir(os.path.join(self.owner.mountPoint, name))
 
@@ -54,8 +54,12 @@ class Implementation(stack.commands.Implementation):
 		else:
 			raise CommandError(self.owner, 'unrecognized pallet format')
 
-		os.system('find . ! -name TRANS.TBL -print | cpio -mpud %s'
-			% new_roll_dir)
+		if not self.owner.dryrun:
+			if not os.path.exists(new_roll_dir):
+				os.makedirs(new_roll_dir)
+			cmd = 'rsync -a --exclude "TRANS.TBL" .  %s/' \
+				% (new_roll_dir)
+			subprocess.call(shlex.split(cmd))
 
 		#
 		# go back to the top of the pallet
@@ -66,5 +70,6 @@ class Implementation(stack.commands.Implementation):
 		# after copying the roll, make sure everyone (apache included)
 		# can traverse the directories
 
-		os.system('find %s -type d -exec chmod a+rx {} \;' % roll_dir)
+		if not self.owner.dryrun:
+			os.system('find %s -type d -exec chmod a+rx {} \;' % roll_dir)
